Changes regarding ga
